componentize ContentIndicator
