+++ components/GalleryElement.vue
@@ -9,10 +9,10 @@
       <span>
         {{ title }}
       </span>
-      <div
+      <ContentIndicator
         v-if="indicator !== undefined"
-        :class="colorCodeClass"
-      ></div>
+        :indicator="indicator"
+      />
     </a>
   </div>
 </template>
@@ -61,8 +61,6 @@
 import Vue from 'vue'
 import VueCompositionAPI, {
   defineComponent,
-  ref,
-  computed,
 } from '@vue/composition-api'
 import Indicator from 'types/Indicator'
 
@@ -87,31 +85,6 @@ export default defineComponent({
       type: Object as () => Indicator,
       default: undefined,
     },
-  },
-  setup(props: any) {
-    const img = ref(props.img)
-    const indicator = ref(props.indicator)
-
-    const imageURL = ref(() => {
-      return `url(${img})`
-    })
-
-    const colorCodeClass = computed(() => {
-      switch (indicator.value) {
-        case 'Mature':
-          return 'color-code cc-mature'
-        case 'Adult':
-          return 'color-code cc-adult'
-        case 'Safe':
-        default:
-          return 'color-code cc-sfw'
-      }
-    })
-
-    return {
-      imageURL,
-      colorCodeClass,
-    }
   },
 })
 </script>
