+++ style/main.sass
@@ -71,20 +71,3 @@ a
   &:hover:after
     visibility: visible
     transform: scaleX(1)
-
-.color-code
-  width: 2rem
-  height: 1rem
-  display: inline-block
-  margin: 0
-  padding: 0
-  border-radius: $mtl-round
-  box-shadow: 0 0 0px 1px white, 0 0 2px 2px black
-  &.cc-mature
-    background-color: rgb(255,159,0)
-    background-image: repeating-linear-gradient(90deg, rgb(255,159,0), rgb(255,159,0) 0.24rem, rgb(200,119,0) 0.25rem, rgb(200,119,0) 0.49rem)
-  &.cc-sfw
-    background-color: rgb(0, 129, 255)
-  &.cc-adult
-    background-color: rgb(155,0,100)
-    background-image: repeating-linear-gradient(49deg, rgb(155,0,100), rgb(155,0,100) 0.24rem, rgb(105,0,50) 0.25rem, rgb(105,0,50) 0.49rem)
